rename propagator file
