+++ src/struphy/models/linear_vlasov_ampere_one_species.py
@@ -20,7 +20,7 @@ from struphy.propagators.base import Propagator
 from struphy.propagators.efield_weights_coupling import EfieldWeightsCoupling
 from struphy.propagators.poisson_solve import PoissonSolve
 from struphy.propagators.push_eta import PushEta
-from struphy.propagators.push_vin_efield import PushVinForceField
+from struphy.propagators.push_v_in_force_field import PushVinForceField
 from struphy.propagators.push_vxb import PushVxB
 
 logger = logging.getLogger("struphy")
