+++ src/struphy/models/linear_vlasov_maxwell_one_species.py
@@ -13,7 +13,7 @@ from struphy.propagators.efield_weights_coupling import EfieldWeightsCoupling
 from struphy.propagators.maxwell_weak_ampere import MaxwellWeakAmpere
 from struphy.propagators.poisson_solve import PoissonSolve
 from struphy.propagators.push_eta import PushEta
-from struphy.propagators.push_vin_efield import PushVinForceField
+from struphy.propagators.push_v_in_force_field import PushVinForceField
 from struphy.propagators.push_vxb import PushVxB
