+++ src/struphy/propagators/__init__.py
@@ -24,7 +24,7 @@ from struphy.propagators.push_eta_pc import PushEtaPC
 from struphy.propagators.push_guiding_center_bx_estar import PushGuidingCenterBxEstar
 from struphy.propagators.push_guiding_center_parallel import PushGuidingCenterParallel
 from struphy.propagators.push_random_diffusion import PushRandomDiffusion
-from struphy.propagators.push_vin_efield import PushVinForceField
+from struphy.propagators.push_v_in_force_field import PushVinForceField
 from struphy.propagators.push_vin_sph_pressure import PushVinSPHpressure
 from struphy.propagators.push_vin_viscous_potential import PushVinViscousPotential
 from struphy.propagators.push_vxb import PushVxB
